+++ lib/ruby2js/converter/next.rb
@@ -5,8 +5,16 @@ module Ruby2JS
     #   (int 1))
 
     handle :next do |n=nil|
-      raise Error.new("next argument #{ n.inspect }", @ast) if n
-      put @next_token.to_s
+      if @next_token == :return
+        put 'return'
+        if n
+          put ' '
+          parse n
+        end
+      else
+        raise Error.new("next argument #{ n.inspect }", @ast) if n
+        put @next_token.to_s
+      end
     end
   end
 end