+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
 * next within a block can return a value
+* handle scans that return zero results with ESLevel < 2020
 
 # 4.0.1 / 2021-02-23