+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # master
 
+# 4.0.2 / 2021-03-02
+
 * next within a block can return a value
 * handle scans that return zero results with ESLevel < 2020
 * redo
 * add rand to filter functions
+* sprockets support
 
 # 4.0.1 / 2021-02-23